Moved HasOneof test into GeneratedClassTest.php, to avoid the extra file.
